gateway: track created services
A new "tracked service" internal API has been added. All calls to
`_create_func` in `ProxyObjectWrapper` now registers the string
reprenstation of the created service as well as the current call
stack to the owning `BlitzGateway` instance. If on deletion, services
are left open, then an ERROR message including the call stack will
be printed.